third_party/blink/web_tests/fast/css/transition-shorthand-cssText-expected.txt

Test for Bug 310403: CssStyleDeclaration improperly sets transition shorthand when transition-duration is set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rule.cssText is '#target { transition-duration: 1s; }'
PASS rule.cssText is '#target { transition-duration: 1s; transition-delay: 0s; }'
PASS rule.cssText is '#target { transition-duration: 1s; transition-delay: 0s; transition-property: all; }'
PASS rule.cssText is '#target { transition: all 1s ease 0s; }'
PASS successfullyParsed is true

TEST COMPLETE
